Inject Satellite Parameters Into Docs

See merge request constellation/constellation!511

+++ docs/conf.py
@@ -1,8 +1,10 @@
 # SPDX-FileCopyrightText: 2024 DESY and the Constellation authors
 # SPDX-License-Identifier: CC0-1.0
 
-import sphinx
 import pathlib
+
+import sphinx
+import sphinx.util.logging
 
 import copy_satellite_docs
 
@@ -108,6 +110,9 @@ myst_fence_as_directive = ["plantuml"]
 myst_enable_extensions = ["colon_fence"]
 myst_update_mathjax = False
 
+# Suppress header warnings from MyST - we check them with markdownlint but cannot disable them in MyST on a per-file level
+suppress_warnings = ["myst.header"]
+
 # breathe settings
 breathe_projects = {
     "Constellation": docsdir.joinpath("doxygen").joinpath("xml"),
@@ -129,29 +134,22 @@ with open("index.md.in", "rt") as index_in, open("index.md", "wt") as index_out:
 
 # Remove existing satellite READMEs
 for path in (docsdir / "satellites").glob("*.md"):
-    path.unlink()
+    if not path.name.startswith("_"):
+        path.unlink()
 
 # Add satellite READMEs to documentation
-satellite_files_cxx = sorted(
-    list((repodir / "cxx" / "satellites").glob("**/README.md"))
-)
-satellite_files_py = sorted(
-    list((repodir / "python" / "constellation" / "satellites").glob("**/README.md"))
-)
+satellite_files_cxx = sorted(list((repodir / "cxx" / "satellites").glob("**/README.md")))
+satellite_files_py = sorted(list((repodir / "python" / "constellation" / "satellites").glob("**/README.md")))
 
 satellites_types_cxx = []
 satellites_types_py = []
 
 for path in satellite_files_cxx:
-    satellite_type = copy_satellite_docs.convert_satellite_readme(
-        path, docsdir / "satellites"
-    )
+    satellite_type = copy_satellite_docs.convert_satellite_readme("cxx", path, docsdir / "satellites")
     satellites_types_cxx.append(f"{satellite_type} <{satellite_type}>")
 
 for path in satellite_files_py:
-    satellite_type = copy_satellite_docs.convert_satellite_readme(
-        path, docsdir / "satellites"
-    )
+    satellite_type = copy_satellite_docs.convert_satellite_readme("py", path, docsdir / "satellites")
     satellites_types_py.append(f"{satellite_type} <{satellite_type}>")
 
 with (
